colorpalettteclient: Fix qmllint warnings
Most of the warnings were straightforward to fix: add missing ids to
lookups, remove an unused import.

There is one interesting case, though: The getCurrentUserImage function
in ColorView was referring "users", which does not exist in that file,
and is not one of its properties. The application however still worked.
That's due to it being found on the root context, as it exists as an id
defined in Main.qml.
It was however already passed down via properties to ColorView, so use
the property instead.
Amends 7a62659880d5cc56316ccf1129e2c694bb138bd7

Change-Id: I6f91dc45302f30d70d8a09b66d0add9b963fa6b2
Reviewed-by: Matthias Rauter <matthias.rauter@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
(cherry picked from commit 78dc366da6c2dff6ce9f4e3a4981d62496b4de2c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>