linguist docs: deprecate //= comments in documentation
The documentation on how to write source code for translation briefly
mentions //= metastrings without going into details. As the
functionality of //= metastrings is not fully implemented and
documented, and is mostly misunderstood and misused, it's deprecated
to use this metastring.

Task-number: QTBUG-139769
Change-Id: If8d1601ffbfeffce924db1b2a996749a4c7ef33a
Reviewed-by: Kai Köhne <kai.koehne@qt.io>
(cherry picked from commit 480007766cb4a9803f03f8586b811cb41e875015)
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>