RobotArm: Make the example less dependent on the Material style
The example directly imports the Material style in various QML files.
If the user wants to run it in a different style, they'd have to modify
it in more places than just the qtquickcontrols2.conf file.

By using file selectors, we can force the application to just force
import the QtQuick.Controls.Material module in files where we're
referencing the Material singleton. This allows us to ignore it in cases
where the user wants to use a different style.

The LabeledSlider would look ugly when using other styles. In order to
improve the look of it, I've used the same trick with file selectors to
alter it for other styles, so that the Slider and Label text are not
overlapping.

Pick-to: 6.10
Change-Id: Ie153097a250d3a646e1c760356937eccf6204a93
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>