Dice: Ignore accelerometer reading until it's ready
The documentation for reading refers to QSensor::reading and it states
that readings are not available immediately after start() is called.
Instead, one should wait for the readingChanged signal. Until then,
reading is documented to be null, so just check for that.

This fixes 'Cannot read property x of null' warnings.

Fixes: QTBUG-140395
Fixes: QTBUG-140385
Pick-to: 6.10
Change-Id: Ibb36e1d8387628eea333cb6d0537dea25f3980b6
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>