Add a topic about vector graphics formats
We have technical and detailed information about specific SVG profiles
and tags that are supported (and equivalent for other formats). This
is useful, but can be hard to process for someone who is only using
a content creation tool and who does not intimately know the internals
of the file format.

So this documentation tries to take a higher level approach to give
graphical designers an idea of what it means that Qt support
"SVG 1.2 Tiny" and how they can adapt their designs to be compatible.

Currently this is written to cover the support we have in Qt 6.10.
And update will be made later for the additions in Qt 6.11.

Pick-to: 6.10
Change-Id: I710121ffaef12514157efa37f60c9cde6e7d80a8
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>