Doc: Remove 'Porting to iOS' documentation page
The parts of the topic that cover Qt Quick apps are outdated. It
implicitly assumes that QML and resources are loaded from the file
system, which is not best practice anymore since a while, on all
platforms. In particular, 'modern' Qt Quick applications that use the
CMake API should not need any of these changes.

Fixes: QTBUG-140067
Pick-to: 6.10
Change-Id: Ib018a6e43c63e8ca88e0baddfa206e6cad3509f0
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>