Doc: Remove the Text Viewer Plugin Example because it is not complete
The Text Viewer Plugin Example does not compile if opened in Qt Creator.
The example was created as a complimentary document for describing the
Text Viewer plugin implementation, but that led to having an uncomplete
example in Qt Creator. This commit removes the example and moves its
content to the Document Viewer demo description.

Fixes: QTBUG-140113
Pick-to: 6.10
Change-Id: Iaf7a4ddbe8d9039bd11ea25413b06bc8129a143f
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>