ToyCustomizer: Make 3D toy bigger in CustomizePage
Currently the toy appears very small in both modes. To make it appear
larger, it's size is increased.

ToyView originally used a GridLayout to position the toy and the
buttons, but this layout placed the toy below the buttons and left
it with too little space. To fix this, the GridLayout was removed
and the components are now positioned using anchors instead.

The toy’s scale has been increased, and because the rabbit’s ears
extend upward, a custom scale is added to it's implementation to
ensure it remains fully visible and well-proportioned.

Fixes: QTBUG-141656
Change-Id: I53bbf37527bec2485326da7a028ce5b27a1c9cdb
Reviewed-by: MohammadHossein Qanbari <mohammad.qanbari@qt.io>