ToyCustomizer: Adjust first and final pages' layout and image sizing
Previously, the "Tab to Start" button overlapped and cropped the 3D
view. This patch moves the button to the bottom of the 3D view by
removing landscape mode, which now shares the same layout as the
portrait mode.

Additionally, this patch updates several images to use sourceSize
instead of implicit sizing for ToyImage items. This prevents image
stretching and blurring.

Fixes: QTBUG-141651
Change-Id: I67ac6be0e62791172c5969f9d12c7bdcd490a705
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>