ToyCustomizer: improve configuration with build and clean steps
After the asset optimization process, it was discovered that some asset
files were unused in the example and therefore removed. Some others
were present in the project but not referenced in any implementation.

Previously, if an asset file was accidentally removed, the build could
fail due to missing references, making it impossible to clean or
rebuild the project. This patch resolves that issue by moving the
asset extraction step to run during every configuration phase, rather
than only when asset downloads are triggered.

Additionally, this patch enhances the clean step to ensure all
extracted files, along with any downloaded compressed files, are
properly removed.

Also, the asset url is updated since it's moved to another server.

Fixes: QTBUG-141235
Change-Id: Id5a3c5b9657659de7b29826e35d4f0403b961af2
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>