Doc: Improve details about Qt Vulnerabilities
Move the section nearer to the issues reporting process.
Explain what a CVE is, and what the role of The Qt Company as a
CNA is. Also list how Qt Framework is identified in a CPE.

While at it, also use Qt Group consistently (except for the CNA
process, where it is actually The Qt Company), and use sentence-style
capitalization for all section headings.

Pick-to: 6.10 6.11
Change-Id: Id62d46ca1fe78c3726d1e4b94032c4c0ca053d77
Reviewed-by: Jerome Pasion <jerome.pasion@qt.io>