Doc: Make titles in in-app purchasing demo more specific
The pages were originally part of the Qt Purchasing module, and
as such are very generically written. This can be confusing to users
that navigate directly to the respective page, as there is no context.

Do a quick fix by
* Adding 'In-app purchasing demo' to each title
* Linking to the example's main page in the first paragraph

While at it, move a somewhat misleading link in the first paragraph
of the Google Play page to the bottom. Also fix grammar error (it's
vs its).

Pick-to: 6.11 6.12
Change-Id: I378b71621458ddd250de6c1dba7e14f628f55c25
Reviewed-by: Alexei Cazacov <alexei.cazacov@qt.io>