ToyCustomizer: Fix the overlapping text issue on OverviewPage
If the name of the selected accessory is too long, the name overlaps
the accessory next to it. There are two reasons for this: first,
the labels did not have any text wrapping enabled; second, OrderItem
had no width constraint, allowing the text to grow beyond the bounds
of its grid cell.

Pick-to: 6.11
Fixes: QTBUG-142693
Change-Id: I50509fbb0eeb02c460ed626e3e14c4de55007d5a
Reviewed-by: MohammadHossein Qanbari <mohammad.qanbari@qt.io>