qtjennydemo: Force lf line-endings for gradlew
Like before, we need to do this for the release packaging to work correctly. In the addition
of observation, this was removed because it was referring to a file that's
no longer there, but the correct change would've been to change it to
force the newlines for the new files.

Pick-to: 6.11 6.10
Change-Id: I1007c9ed4771cb68fa15bcdbad53de639b6866ae
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>